# docs/swagger.yaml
info:
  contact: {}
  license: {}
paths:
  /api/v1/articles:
    get:
      parameters:
      - description: Title
        in: query
        name: title
        type: string
      produces:
      - application/json
      summary: 获取所有文章
      tags:
      - 文章模块
    post:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Title
        in: query
        name: title
        required: true
        type: string
      - description: Desc
        in: query
        name: desc
        type: string
      - description: Content
        in: query
        name: content
        type: string
      - description: CreatedBy
        in: query
        name: created_by
        type: string
      - description: tagId
        in: query
        name: tag_id
        type: integer
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      summary: 添加文章
      tags:
      - 文章模块
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      summary: 删除文章标签
      tags:
      - 文章模块
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      summary: 获取指定文章
      tags:
      - 文章模块
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Title
        in: query
        name: title
        required: true
        type: string
      - description: Desc
        in: query
        name: desc
        type: string
      - description: Content
        in: query
        name: content
        type: string
      - description: ModifiedBy
        in: query
        name: modified_by
        type: string
      - description: TagId
        in: query
        name: tag_id
        type: integer
      - description: State
        in: query
        name: state
        type: integer
      produces:
      - application/json
      summary: 修改文章标签
      tags:
      - 文章模块
  /api/v1/tags:
    get:
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      produces:
      - application/json
      summary: 获取所有标签
      tags:
      - 文章模块
    post:
      parameters:
      - description: Name
        in: query
        name: name
        required: true
        type: string
      - description: State
        in: query
        name: state
        type: integer
      - description: CreatedBy
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      summary: 新增文章标签
      tags:
      - 文章模块
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      summary: 删除文章标签
      tags:
      - 文章模块
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: ID
        in: query
        name: name
        required: true
        type: string
      - description: ModifiedBy
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      summary: 修改文章标签
      tags:
      - 文章模块
  /auth:
    get:
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      - description: Password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"ok", "data":{}}'
      summary: 获取JwtToken
      tags:
      - 用户模块
swagger: "2.0"